Add pipeline setup functions
The `NewExportPipeline` function configures an Exporter from environment
variables and build configurable trace and metric pipelines with the
exporter. The built `trace.Provider` and metric `Controller` are
returned.

The `InstallNewPipeline` function configures an Exporter from
environment variables and installs it into a telemetry pipeline using
default configuration as well as setting it as the global `*Provider`.

The examples for the package are updated to show how these functions can
be used.